split protocol module into multiple files
