+++ src/protocol/mod.rs
@@ -1,0 +1,5 @@
+mod connection;
+mod network_config;
+
+pub use connection::Connection;
+pub use network_config::NetworkConfig;
